requirements.txt
# apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
# Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
# command-not-found==0.3
cryptography==3.4.8
# cupshelpers==1.0
# dbus-python==1.2.18
# defer==1.0.6
distlib==0.3.8
distro==1.7.0
# distro-info==1.1+ubuntu0.1
Django==5.0.4
django-allauth==0.62.1
django-user-agents==0.4.0
# duplicity==0.8.21
fasteners==0.14.1
filelock==3.15.1
future==0.18.2
# gyp==0.1
httplib2==0.20.2
idna==2.5
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
# louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
protobuf==3.12.4
ptyprocess==0.7.0
# pycairo==1.20.1
pycparser==2.22
# pycups==2.0.1
# PyGObject==3.42.1
psycopg2-binary==2.9.9
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
# python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43
pytz==2022.1
pyxdg==0.27
# PyYAML==5.4.1
# reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
# systemd-python==234
typing_extensions==4.11.0
# ubuntu-advantage-tools==8001
# ubuntu-drivers-common==0.0.0
# ufw==0.36.1
# unattended-upgrades==0.1
urllib3==1.26.5
# usb-creator==0.3.7
virtualenv==20.26.2
wadllib==1.3.6
xdg==5
# xkit==0.0.0
zipp==1.0.0
